Refactor: libcrmservice: isolate nagios-related code in its own source file
... for better code isolation and symmetry with other resource types
