+++ lib/services/Makefile.am
@@ -1,5 +1,5 @@
 #
-# Copyright 2012-2018 the Pacemaker project contributors
+# Copyright 2012-2019 the Pacemaker project contributors
 #
 # The version control history for this file may have further details.
 #
@@ -13,7 +13,8 @@ AM_CPPFLAGS			= -I$(top_srcdir)/include
 
 lib_LTLIBRARIES			= libcrmservice.la
 noinst_HEADERS			= pcmk-dbus.h upstart.h systemd.h \
-				  services_lsb.h services_private.h
+				  services_lsb.h services_nagios.h \
+				  services_private.h
 
 libcrmservice_la_LDFLAGS	= -version-info 30:0:2
 libcrmservice_la_CPPFLAGS	= -DOCF_ROOT_DIR=\"@OCF_ROOT_DIR@\" $(AM_CPPFLAGS)
@@ -24,7 +25,9 @@ libcrmservice_la_LDFLAGS	+= $(LDFLAGS_HARDENED_LIB)
 
 libcrmservice_la_LIBADD		= $(top_builddir)/lib/common/libcrmcommon.la $(DBUS_LIBS)
 
-libcrmservice_la_SOURCES	= services.c services_linux.c services_lsb.c
+libcrmservice_la_SOURCES	= services.c
+libcrmservice_la_SOURCES	+= services_linux.c
+libcrmservice_la_SOURCES	+= services_lsb.c
 if BUILD_DBUS
 libcrmservice_la_SOURCES	+= dbus.c
 endif
@@ -33,4 +36,7 @@ libcrmservice_la_SOURCES	+= upstart.c
 endif
 if BUILD_SYSTEMD
 libcrmservice_la_SOURCES	+= systemd.c
+endif
+if BUILD_NAGIOS
+libcrmservice_la_SOURCES	+= services_nagios.c
 endif
