+++ lib/services/services_linux.c
@@ -1188,50 +1188,3 @@ services__ocf_agent_exists(const char *provider, const char *agent)
     free(buf);
     return rc;
 }
-
-#if SUPPORT_NAGIOS
-GList *
-resources_os_list_nagios_agents(void)
-{
-    GList *plugin_list = NULL;
-    GList *result = NULL;
-    GList *gIter = NULL;
-
-    plugin_list = get_directory_list(NAGIOS_PLUGIN_DIR, TRUE, TRUE);
-
-    /* Make sure both the plugin and its metadata exist */
-    for (gIter = plugin_list; gIter != NULL; gIter = gIter->next) {
-        const char *plugin = gIter->data;
-        char *metadata = crm_strdup_printf(NAGIOS_METADATA_DIR "/%s.xml", plugin);
-        struct stat st;
-
-        if (stat(metadata, &st) == 0) {
-            result = g_list_append(result, strdup(plugin));
-        }
-
-        free(metadata);
-    }
-    g_list_free_full(plugin_list, free);
-    return result;
-}
-
-gboolean
-services__nagios_agent_exists(const char *name)
-{
-    char *buf = NULL;
-    gboolean rc = FALSE;
-    struct stat st;
-
-    if (name == NULL) {
-        return rc;
-    }
-
-    buf = crm_strdup_printf(NAGIOS_PLUGIN_DIR "/%s", name);
-    if (stat(buf, &st) == 0) {
-        rc = TRUE;
-    }
-
-    free(buf);
-    return rc;
-}
-#endif
